payment

+++ database/migrations/1628934263031_payments.ts
@@ -13,7 +13,7 @@ export default class Payments extends BaseSchema {
       table.decimal('amount', 9, 2)
       table.string('nickname', 20)
       table.string('operator', 7)
-      table.integer('kassaPaymentId')
+      table.bigInteger('kassaPaymentId')
       table.string('email', 30)
       table.integer('phone', 15)
       table.integer('currencyId', 2)
